Variables now work in blocks
I'm still not sure whether this is the best implementation tho. Maybe it'll have to be changed in the future. It could probably be simplified a lot. I'm not really proud of it, but it works. Same with the blocks method hooking system. I should look into improving this at some point, but for now it works and that's important.